backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==4.2.16
django-cors-headers==4.3.1
django-filter==24.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.2.3
drf-yasg==1.21.7
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
isort==5.13.2
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
